fixes affiliate codes
